Move DB accesses from constructor to runtime (#634)
DB accesses in constructors are discouraged by the ATC framework and can lead to problems in upgrade shadow systems.

This moves the check for a profile for the current user from the constructor to QUERY_ATTRIBUTES. This means the checks no longer display as having no settings in SCI when a profile is active for the current user, but instead show a corresponding message when the settings button is clicked.